+++ be/demo/src/main/java/com/example/demo/cart/domain/Cart.java
@@ -7,7 +7,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
-import javax.persistence.Table;
 
 import com.example.demo.pay.domain.Payment;
 
@@ -17,8 +16,8 @@ import lombok.Getter;
 public class Cart {
    @Id 
    @GeneratedValue(strategy = GenerationType.IDENTITY)
-   @Column(name="cart_no") private int cartNo;
-   @Column(name="cart_amount") private int cartAmount;
+   @Column(name="cart_no") private long cartNo;
+   @Column(name="cart_amount") private long cartAmount;
    
    @OneToOne
    @JoinColumn(name = "pay_no")
