+++ be/demo/src/main/java/com/example/demo/lvl/domain/Level.java
@@ -3,7 +3,6 @@ package com.example.demo.lvl.domain;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.Table;
 
 import lombok.Getter;
 
@@ -13,6 +12,6 @@ public class Level {
 
 	@Id
 	@Column(name = "level")
-	private int level;
+	private long level;
 
 }
