+++ be/demo/src/main/java/com/example/demo/lvl/service/LevelServiceImpl.java
@@ -17,13 +17,13 @@ public class LevelServiceImpl extends AbstractService<Level>
 								implements LevelService{
 	private final LevelRepository repo;
 
-	@Override public int save(Level t) {return (repo.save(t)!=null) ? 1 : 0 ;}
-	@Override public int count() {return (int) repo.count();}
-	@Override public Level getOne(int id) {return repo.getOne(id);}
-	@Override public Optional<Level> findById(int id) {return repo.findById(id);}
-	@Override public boolean existsById(int id) {return repo.existsById(id);}
+	@Override public long save(Level t) {return (repo.save(t)!=null) ? 1 : 0 ;}
+	@Override public long count() {return (long) repo.count();}
+	@Override public Level getOne(long id) {return repo.getOne(id);}
+	@Override public Optional<Level> findById(long id) {return repo.findById(id);}
+	@Override public boolean existsById(long id) {return repo.existsById(id);}
 	@Override public List<Level> findAll() {return repo.findAll();}
-	@Override public int delete(Level t) {
+	@Override public long delete(Level t) {
 		repo.delete(t); 
 		return (getOne(t.getLevel())==null) ? 1 : 0;
 	}
