+++ be/demo/src/main/java/com/example/demo/pay/domain/Payment.java
@@ -23,9 +23,9 @@ import lombok.Getter;
 
 @Entity @Getter
 public class Payment {
-   @Id @GeneratedValue(strategy = GenerationType.IDENTITY) @Column(name="pay_no") private int payNo;
+   @Id @GeneratedValue(strategy = GenerationType.IDENTITY) @Column(name="pay_no") private long payNo;
    @Column(name="pay_price") private String payPrice;
-   @Column(name="pay_amount") private int payAmount;
+   @Column(name="pay_amount") private long payAmount;
    @Column(name="dvr_fee") private String dvrFee;
    @Column(name="pay_date") private String payDate;
    @Column(name="pay_state") private String payState;
