+++ be/demo/src/main/java/com/example/demo/rcv/domain/Receiver.java
@@ -8,7 +8,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 import com.example.demo.pay.domain.Payment;
 
@@ -18,7 +17,7 @@ import lombok.Getter;
 public class Receiver {
       @Id 
       @GeneratedValue(strategy = GenerationType.IDENTITY)
-      @Column(name="rcv_no") private int rcvNo;
+      @Column(name="rcv_no") private long rcvNo;
       @Column(name="rcv_name") private String rcvName;
       @Column(name="rcv_phone") private String rcvPhone;
       @Column(name="rcv_addr") private String rcvAddr;
