+++ be/demo/src/main/java/com/example/demo/rpl/controller/ReplyController.java
@@ -27,32 +27,32 @@ public class ReplyController extends AbstractController<Reply> {
 	final ReplyServiceImpl service;
 	
 	@PostMapping("/save")
-	public ResponseEntity<Integer> save(@RequestBody Reply t) {
+	public ResponseEntity<Long> save(@RequestBody Reply t) {
 		return ResponseEntity.ok(service.save(t));
 	}
 
 	@DeleteMapping("/delete")
-	public ResponseEntity<Integer> delete(@RequestBody Reply t) {
+	public ResponseEntity<Long> delete(@RequestBody Reply t) {
 		return ResponseEntity.ok(service.delete(t));
 	}
 
 	@GetMapping("/count")
-	public ResponseEntity<Integer> count() {
+	public ResponseEntity<Long> count() {
 		return ResponseEntity.ok(service.count());
 	}
 
 	@GetMapping("/one/{id}")
-	public ResponseEntity<Reply> getOne(@PathVariable int id) {
+	public ResponseEntity<Reply> getOne(@PathVariable long id) {
 		return ResponseEntity.ok(service.getOne(id));
 	}
 
 	@GetMapping("/find/{id}")
-	public ResponseEntity<Optional<Reply>> findById(@PathVariable int id) {
+	public ResponseEntity<Optional<Reply>> findById(@PathVariable long id) {
 		return ResponseEntity.ok(service.findById(id));
 	}
 
 	@GetMapping("/exists/{id}")
-	public ResponseEntity<Boolean> existsById(@PathVariable int id) {
+	public ResponseEntity<Boolean> existsById(@PathVariable long id) {
 		return ResponseEntity.ok(service.existsById(id));
 	}
