+++ be/demo/src/main/java/com/example/demo/rpl/domain/Reply.java
@@ -7,7 +7,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 
 import com.example.demo.brd.domain.Board;
 import com.example.demo.usr.domain.User;
@@ -16,7 +15,7 @@ import lombok.Getter;
 
 @Entity @Getter
 public class Reply {
-	@Id @Column(name="rpl_no") @GeneratedValue(strategy = GenerationType.IDENTITY) private int rplNo;
+	@Id @Column(name="rpl_no") @GeneratedValue(strategy = GenerationType.IDENTITY) private long rplNo;
 	@Column(name="rpl_content") private String rplContent;
