+++ be/demo/src/main/java/com/example/demo/rpl/domain/ReplyDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 
 @Data @Lazy @Component
 public class ReplyDto {
-	private int rplNo;
+	private long rplNo;
 	private String rplContent;
 
 }
