+++ be/demo/src/main/java/com/example/demo/rpl/service/ReplyServiceImpl.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Service;
 import com.example.demo.cmm.service.AbstractService;
 import com.example.demo.rpl.domain.Reply;
 import com.example.demo.rpl.repository.ReplyRepository;
-import com.example.demo.rpl.service.ReplyService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -18,13 +17,13 @@ public class ReplyServiceImpl extends AbstractService<Reply>
 								implements ReplyService{
 	private final ReplyRepository repo;
 
-	@Override public int save(Reply t) {return (repo.save(t)!=null) ? 1 : 0 ;}
-	@Override public int count() {return (int) repo.count();}
-	@Override public Reply getOne(int id) {return repo.getOne(id);}
-	@Override public Optional<Reply> findById(int id) {return repo.findById(id);}
-	@Override public boolean existsById(int id) {return repo.existsById(id);}
+	@Override public long save(Reply t) {return (repo.save(t)!=null) ? 1 : 0 ;}
+	@Override public long count() {return (long) repo.count();}
+	@Override public Reply getOne(long id) {return repo.getOne(id);}
+	@Override public Optional<Reply> findById(long id) {return repo.findById(id);}
+	@Override public boolean existsById(long id) {return repo.existsById(id);}
 	@Override public List<Reply> findAll() {return repo.findAll();}
-	@Override public int delete(Reply t) {
+	@Override public long delete(Reply t) {
 		repo.delete(t); 
 		return (getOne(t.getRplNo())==null) ? 1 : 0;
 	}
